Refactor application into components
