+++ src/ModelManager.h
@@ -1,0 +1,21 @@
+class ModelManager {
+public:
+    int LoadModel(const std::string &modelPath);
+    void UnloadModel(int index);
+
+    size_t Count() const { return models_.size(); }
+    Model* GetModel(int index) { return index>=0 && index<(int)models_.size()? models_[index].get():nullptr; }
+    Shader* GetShader(int index) { return index>=0 && index<(int)shaders_.size()? shaders_[index].get():nullptr; }
+    Transform* GetTransform(int index) { return index>=0 && index<(int)transforms_.size()? transforms_[index].get():nullptr; }
+    glm::vec3 GetDimensions(int index) const { return meshDimensions_[index]; }
+    const std::string &GetPath(int index) const { return modelPaths_[index]; }
+
+    void EnforceGridConstraint(int index);
+
+private:
+    std::vector<std::unique_ptr<Shader>> shaders_;
+    std::vector<std::unique_ptr<Model>> models_;
+    std::vector<std::unique_ptr<Transform>> transforms_;
+    std::vector<glm::vec3> meshDimensions_;
+    std::vector<std::string> modelPaths_;
+};
